Move built-in signal handler to dedicated file
