+++ wagtail_bucketav/__init__.py
@@ -1,4 +1,8 @@
+import logging
+
 __version__ = "0.0.1"
+
+logger = logging.getLogger("__name__")
 
 
 def get_version():
